Add simple RabbitMQ appender
Add RabbitMQ appender (with no real features) that just writes to an
exchange instead of to a specific queue.

Future would would merge this into the GelfAmqpAppender, providing more
specific semantics for choosing whether you want to write to an exchange
or queue, what type the exchange might be, whether you want to create
the exchange or the queue.